% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mylm.R
\name{myGeneric}
\alias{myGeneric}
\title{Neue generische Funktion}
\usage{
myGeneric(x, ...)
}
\arguments{
\item{x}{ein object}

\item{...}{zusaetzliche Argumente, die an Methoden weitergereicht werden}
}
\description{
Test
}

