% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mylm.R
\name{mylm}
\alias{mylm}
\title{Mein lineares Modell}
\usage{
mylm(y, X)
}
\arguments{
\item{y}{abhängige Variable}

\item{X}{Design-Matrix}
}
\value{
Rueckgabewert
}
\description{
Berechnet die Koeffizienten eines linearen Modells
}
\details{
Mehr Details
}
\examples{
mylm(rnorm(10), rnorm(10))
}

